requirements.txt
click==8.2.1
spotipy==2.25.1
python-dotenv==1.1.1
ytmusicapi==1.10.3